Fixed bugs

+++ app/src/main/java/jhu/voiceit/LoginActivity.java
@@ -29,7 +29,7 @@ public class LoginActivity extends AppCompatActivity {
     private ImageButton regis;
 
     private String username;
-    private int numPosts;
+    private long numPosts;
     private EditText usrenter;
     private EditText passenter;
 
@@ -155,7 +155,7 @@ public class LoginActivity extends AppCompatActivity {
                     @Override
                     public void onDataChange(DataSnapshot dataSnapshot) {
                         username = (String) dataSnapshot.child("username").getValue();
-                        numPosts = (int) dataSnapshot.child("numPosts").getValue();
+                        numPosts = (long) dataSnapshot.child("numPosts").getValue();
                     }
 
                     @Override
@@ -167,7 +167,7 @@ public class LoginActivity extends AppCompatActivity {
                 SharedPreferences.Editor peditor = myPrefs.edit();
                 peditor.putString("Email",usrenter.getText().toString());
                 peditor.putString("UserName", username);
-                peditor.putInt("numPosts", numPosts);
+                peditor.putLong("numPosts", numPosts);
                 peditor.putString("auth_token", authData.getToken());
                 peditor.putString("UID", userId);
                 peditor.commit();
